added configs for player weapons

+++ Assets/_Game/Scripts/Components/Weapons/LightningShooterComponent.cs
@@ -6,6 +6,6 @@ public class LightningShooterComponent : IComponent
     public float Cooldown;
     public float EffectRadius;
     public int TargetDamage;
-    public int AOEDamage;
+    public float AOEDamage;
     public float StunDuration;
 }
